+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Current Trunk
   function references spec.
 - HeapType::ext has been restored but is no longer a subtype of HeapType::any to
   match the latest updates in the GC spec. (#4898)
+- `i31ref` and `dataref` are now nullable. (TODO: PR number)
 
 v109
 ----